+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
+## [Unreleased]
+#### Fixed
+- Trailing digits in amount input field
+
+#### Added
+- Error message when input amount is less than min amount
+
 ## [1.11.0-rc.0] - 2020-02-05
 #### Fixed
 - A bug with color scheme